--- .babelrc
@@ -1,13 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      { "targets": { "browsers": [">0.25%", "not dead"] } }
-    ],
-    "@babel/preset-react"
-  ],
-  "plugins": [
-    "@babel/plugin-proposal-optional-chaining",
-    "@babel/plugin-proposal-class-properties"
-  ]
-}
