--- src/components/Seo.tsx
@@ -1,87 +1,0 @@
-/**
- * SEO component that queries for data with
- *  Gatsby's useStaticQuery React hook
- *
- * See: https://www.gatsbyjs.org/docs/use-static-query/
- */
-
-import React from "react";
-import Helmet from "react-helmet";
-import { useStaticQuery, graphql, withPrefix } from "gatsby";
-
-export default () => {
-  const {
-    site,
-  }: {
-    site: {
-      siteMetadata: {
-        title: string;
-        description: string;
-        author: string;
-      };
-    };
-  } = useStaticQuery(
-    graphql`
-      query {
-        site {
-          siteMetadata {
-            title
-            description
-            author
-          }
-        }
-      }
-    `
-  );
-
-  const metaDescription = site.siteMetadata.description;
-
-  return (
-    <Helmet
-      htmlAttributes={{
-        lang: "ja",
-      }}
-      title={site.siteMetadata.title}
-      meta={[
-        {
-          name: `description`,
-          content: metaDescription,
-        },
-        {
-          property: `og:title`,
-          content: site.siteMetadata.title,
-        },
-        {
-          property: `og:description`,
-          content: metaDescription,
-        },
-        {
-          property: `og:type`,
-          content: `website`,
-        },
-        {
-          name: `twitter:card`,
-          content: `summary`,
-        },
-        {
-          name: `twitter:creator`,
-          content: site.siteMetadata.author,
-        },
-        {
-          name: `twitter:title`,
-          content: site.siteMetadata.title,
-        },
-        {
-          name: `twitter:description`,
-          content: metaDescription,
-        },
-      ]}
-    >
-      <meta charSet="utf-8" />
-      <link rel="stylesheet" href={withPrefix("css/bootstrap.min.css")} />
-      <link rel="stylesheet" href={withPrefix("css/index.css")} />
-      <link rel="stylesheet" href={withPrefix("css/mobile.css")} />
-    </Helmet>
-  );
-  1;
-};
