+++ src/container/MainContent/MainContentContainer.tsx
@@ -1,31 +1,14 @@
-import React, { useRef, useEffect, useState } from "react";
+import React, { useRef, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { withPrefix, useStaticQuery, graphql } from "gatsby";
 
 import PlayAudioState from "../../state/PlayAudioState";
 import ButtonSectionContainer from "../ButtonSectionContainer/ButtonSectionContainer";
 
 import AppState from "../../state/AppState";
-import { StopAudioAction, FetchVoiceListAction } from "../../actions/Actions";
-import VoiceDataQueryResult from "../../model/GraphQLResult/VoiceData";
+import { StopAudioAction } from "../../actions/Actions";
 
 export default () => {
   const [isFirstPlay, setIsFirstPlay] = useState(true);
-
-  const data: VoiceDataQueryResult = useStaticQuery(
-    graphql`
-      query {
-        allFile(sort: { fields: [relativePath], order: ASC }) {
-          edges {
-            node {
-              name
-              relativePath
-            }
-          }
-        }
-      }
-    `
-  );
   const dispatch = useDispatch();
   const audioRef = useRef<HTMLAudioElement>();
   const {
@@ -36,11 +19,6 @@ export default () => {
     playAudio: PlayAudioState;
   } = useSelector(({ app, playAudio }) => ({ app, playAudio }));
 
-  useEffect(() => {
-    if (app.voiceList.length === 0) {
-      dispatch(FetchVoiceListAction(data));
-    }
-  });
   return (
     <main
       onClick={() => {
@@ -64,7 +42,7 @@ export default () => {
         ref={audioRef}
         onCanPlay={() => audioRef.current.play()}
         onEnded={() => dispatch(StopAudioAction())}
-        src={withPrefix(`audio/${playAudio.filename}`)}
+        src={`static/audio/${playAudio.filename}`}
         muted
         autoPlay
       />
